// src/gen/zod.gen.ts
export const zInsertTextRequest = z.object({
    text: z.string().min(1),
    file_source: z.string().min(0).nullish()
});